requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-groq>=0.1.0
pypdf>=4.0.0
python-dotenv>=1.0.1
streamlit>=1.38.0
pytest>=8.3.0
requests>=2.32.0
langsmith>=0.1.105
groq>=0.11.0
sentence-transformers>=3.0.0
qdrant-client>=1.7.0
qdrant-client>=1.7.0

